+++ src/bika/lims/content/contact.py
@@ -21,7 +21,6 @@
 import types
 
 from AccessControl import ClassSecurityInfo
-from Acquisition import aq_base
 from Acquisition import aq_inner
 from Acquisition import aq_parent
 from bika.lims import bikaMessageFactory as _
@@ -41,8 +40,6 @@ from Products.CMFCore.permissions import ModifyPortalContent
 from Products.CMFPlone.utils import safe_unicode
 from senaite.core.p3compat import cmp
 from zope.interface import implements
-
-ACTIVE_STATES = ["active"]
 
 
 schema = Person.schema.copy() + atapi.Schema((
@@ -247,9 +244,8 @@ class Contact(Person):
         # N.B. Local owner role and client group applies only to client
         #      contacts, but not lab contacts.
         if IClient.providedBy(self.aq_parent):
-            # Add user to "Clients" group
-            self._addUserToGroup(username, group="Clients")
-            self._recursive_reindex_object(self.aq_parent)
+            # add user to clients group
+            self.aq_parent.add_user_to_group(username)
 
         return True
 
@@ -284,37 +280,10 @@ class Contact(Person):
         # N.B. Local owner role and client group applies only to client
         #      contacts, but not lab contacts.
         if IClient.providedBy(self.aq_parent):
-            # Remove user from "Clients" group
-            self._delUserFromGroup(username, group="Clients")
-            self._recursive_reindex_object(self.aq_parent)
+            # remove user from clients group
+            self.aq_parent.del_user_from_group(username)
 
         return True
-
-    @security.private
-    def _addUserToGroup(self, username, group="Clients"):
-        """Add user to the goup
-        """
-        portal_groups = api.portal.get_tool("portal_groups")
-        group = portal_groups.getGroupById(group)
-        group.addMember(username)
-
-    @security.private
-    def _delUserFromGroup(self, username, group="Clients"):
-        """Remove user from the group
-        """
-        portal_groups = api.portal.get_tool("portal_groups")
-        group = portal_groups.getGroupById(group)
-        group.removeMember(username)
-
-    def _recursive_reindex_object(self, obj):
-        """Reindex object after user linking
-        """
-        if hasattr(aq_base(obj), "objectValues"):
-            for child_obj in obj.objectValues():
-                self._recursive_reindex_object(child_obj)
-
-        logger.debug("Reindexing object {}".format(repr(obj)))
-        obj.reindexObject(idxs=["allowedRolesAndUsers"])
 
 
 atapi.registerType(Contact, PROJECTNAME)