+++ src/senaite/core/upgrade/v02_05_000.py
@@ -18,6 +18,8 @@
 # Copyright 2018-2023 by it's authors.
 # Some rights reserved, see README and LICENSE.
 
+import transaction
+from Acquisition import aq_base
 from bika.lims import api
 from senaite.core import logger
 from senaite.core.api.catalog import add_index
@@ -28,6 +30,8 @@ from senaite.core.catalog import CLIENT_CATALOG
 from senaite.core.catalog import REPORT_CATALOG
 from senaite.core.catalog import SAMPLE_CATALOG
 from senaite.core.config import PROJECTNAME as product
+from senaite.core.permissions import ManageBika
+from senaite.core.registry import get_registry_record
 from senaite.core.setuphandlers import add_dexterity_items
 from senaite.core.setuphandlers import setup_catalog_mappings
 from senaite.core.setuphandlers import setup_core_catalogs
@@ -37,6 +41,25 @@ from senaite.core.upgrade.utils import uncatalog_brain
 
 version = "2.5.0"  # Remember version number in metadata.xml and setup.py
 profile = "profile-{0}:default".format(product)
+
+CONTENT_ACTIONS = [
+    # portal_type, action
+    ("Client", {
+        "id": "manage_access",
+        "name": "Manage Access",
+        "action": "string:${object_url}/@@sharing",
+        # NOTE: We use this permission to hide the action from logged in client
+        # contacts
+        "permission": ManageBika,
+        # "permission": "Sharing page: Delegate roles",
+        "category": "object",
+        "visible": True,
+        "icon_expr": "",
+        "link_target": "",
+        "condition": "",
+        "insert_after": "edit",
+    }),
+]
 
 
 @upgradestep(product, version)
@@ -135,7 +158,7 @@ def setup_catalogs(tool):
     logger.info("Setup Catalogs [DONE]")
 
 
-def update_report_catalog(self):
+def update_report_catalog(tool):
     """Update indexes in report catalog and add new metadata columns
     """
     logger.info("Update report catalog ...")
@@ -150,3 +173,123 @@ def update_report_catalog(self):
     del_column(REPORT_CATALOG, "getDatePrinted")
 
     logger.info("Update report catalog [DONE]")
+
+
+def import_registry(tool):
+    """Import registry step from profiles
+    """
+    portal = tool.aq_inner.aq_parent
+    setup = portal.portal_setup
+    setup.runImportStepFromProfile(profile, "plone.app.registry")
+
+
+def create_client_groups(tool):
+    """Create for all Clients an explicit Group
+    """
+    logger.info("Create client groups ...")
+    clients = api.search({"portal_type": "Client"}, CLIENT_CATALOG)
+    total = len(clients)
+    for num, client in enumerate(clients):
+        obj = api.get_object(client)
+        logger.info("Processing client %s/%s: %s"
+                    % (num+1, total, obj.getName()))
+
+        # recreate the group
+        obj.remove_group()
+
+        # skip group creation
+        if not get_registry_record("auto_create_client_group", True):
+            logger.info("Auto group creation is disabled in registry. "
+                        "Skipping group creation ...")
+            continue
+
+        group = obj.create_group()
+        # add all linked client contacts to the group
+        for contact in obj.getContacts():
+            user = contact.getUser()
+            if not user:
+                continue
+            logger.info("Adding user '%s' to the client group '%s'"
+                        % (user.getId(), group.getId()))
+            obj.add_user_to_group(user)
+
+    logger.info("Create client groups [DONE]")
+
+
+def reindex_client_security(tool):
+    """Reindex client object security to grant the owner role for the client
+       group to all contents
+    """
+    logger.info("Reindex client security ...")
+
+    clients = api.search({"portal_type": "Client"}, CLIENT_CATALOG)
+    total = len(clients)
+    for num, client in enumerate(clients):
+        obj = api.get_object(client)
+        logger.info("Processing client %s/%s: %s"
+                    % (num+1, total, obj.getName()))
+
+        if not obj.get_group():
+            logger.info("No client group exists for client %s. "
+                        "Skipping reindexing ..." % obj.getName())
+            continue
+
+        _recursive_reindex_object_security(obj)
+
+        logger.info("Commiting client %s/%s" % (num+1, total))
+        transaction.commit()
+        logger.info("Commit done")
+
+        # Flush the object from memory
+        obj._p_deactivate()
+
+    logger.info("Reindex client security [DONE]")
+
+
+def _recursive_reindex_object_security(obj):
+    """Recursively reindex object security for the given object
+    """
+    if hasattr(aq_base(obj), "objectValues"):
+        for child_obj in obj.objectValues():
+            _recursive_reindex_object_security(child_obj)
+    obj.reindexObject(idxs=["allowedRolesAndUsers"])
+    obj._p_deactivate()
+
+
+def add_content_actions(tool):
+    logger.info("Add cotent actions ...")
+    portal_types = api.get_tool("portal_types")
+    for record in CONTENT_ACTIONS:
+        portal_type, action = record
+        type_info = portal_types.getTypeInfo(portal_type)
+        action_id = action.get("id")
+        # remove any previous added actions with the same ID
+        _remove_action(type_info, action_id)
+        # pop out the position info
+        insert_after = action.pop("insert_after", None)
+        # add the action
+        type_info.addAction(**action)
+        # sort the action to the right position
+        actions = type_info._cloneActions()
+        action_ids = map(lambda a: a.id, actions)
+        if insert_after in action_ids:
+            ref_index = action_ids.index(insert_after)
+            index = action_ids.index(action_id)
+            action = actions.pop(index)
+            actions.insert(ref_index + 1, action)
+            type_info._actions = tuple(actions)
+
+        logger.info("Added action id '%s' to '%s'",
+                    action_id, portal_type)
+    logger.info("Add content actions [DONE]")
+
+
+def _remove_action(type_info, action_id):
+    """Removes the action id from the type passed in
+    """
+    actions = map(lambda action: action.id, type_info._actions)
+    if action_id not in actions:
+        return True
+    index = actions.index(action_id)
+    type_info.deleteActions([index])
+    return _remove_action(type_info, action_id)